refactor gemini_utils   - unify the method for sending all kinds of attachement to gemini chat   - use built in consts for the different mime types   - read json files as text for prompt tat includes json files.. this makes get_response_from_gemini async   - change all calling method accordingly remove get_mime_type from file_utils

+++ src/commands_automator/llm/llm_agents/gemini/gemini_utils.py
@@ -1,68 +1,109 @@
 import io
 import logging
+import mimetypes
 from google import genai
 from google.genai.chats import Chat
 from google.genai.types import FileData
+from google.genai.types import File
 from google.genai.types import Part
 from PIL import Image
+
+from commands_automator.utils import file_utils
 
 class GeminiUtils:
     GEMINI_MODEL = "gemini-2.5-flash"
 
     CONFIG_RESPONSE_MIME_TYPE = "response_mime_type"
-    MIME_TYPE_JSON = "application/json"
-    MIME_TYPE_TEXT = "text/plain"
 
     def __init__(self, gemini_client: genai.Client):
         self.gemini_client: genai.Client = gemini_client
+        self.delete_all_files()
 
     def init_chat(self) -> Chat:
-        config = {
+        self.config = {
             "temperature": 0,   # Creativity (0: deterministic, 1: high variety)
             "top_p": 0.95,       # Focus on high-probability words
             "top_k": 64,        # Consider top-k words for each step
             "max_output_tokens": 8192,  # Limit response length
-            self.CONFIG_RESPONSE_MIME_TYPE: self.MIME_TYPE_JSON,
+            self.CONFIG_RESPONSE_MIME_TYPE: mimetypes.types_map['.json'],
         }
-        chat  = self.gemini_client.chats.create(
+        chat = self.gemini_client.chats.create(
             model= self.GEMINI_MODEL,
-            config=config,
+            config=self.config,
             history=[]
         )
 
         return chat
-    
-    def get_response_from_gemini(self, prompt: str,  chat: Chat, base64_decoded : str = None, file_paths : list[str] = None) -> str:
-            try:
-                chat._config[self.CONFIG_RESPONSE_MIME_TYPE] = self.MIME_TYPE_TEXT           
-                if base64_decoded:
-                    image = Image.open(io.BytesIO(base64_decoded))
-                    gemini_response = chat.send_message([prompt, image])
-                elif file_paths:
-                    gemini_text = ""
-                    for file in file_paths:
-                        file_part = self.upload_large_file_to_google_cloud(file,  self.MIME_TYPE_TEXT)
-                        gemini_response = chat.send_message(file_part)
-                        gemini_text = f"{gemini_text}\n{gemini_response.text}"
-                    gemini_response = chat.send_message(prompt)
-                    gemini_text = f"{gemini_text}\n{gemini_response.text}"
-                    return gemini_text
-                else:
-                    gemini_response = chat.send_message(prompt)                   
-                gemini_text = gemini_response.text
-                return gemini_text
-            except Exception as e:
-                logging.error(f"Error using Gemini {e}", exc_info=True)
-                return "Sorry, I couldn't process your request with Gemini or MCP tools."
 
-    def upload_large_file_to_google_cloud(self, file_path, mime_type) -> Part:
-        if not mime_type:
-            raise ValueError("mime_type parameter is required")
-        logging.debug("got a file to upload")
-        file = self.gemini_client.files.upload(file=file_path)        
-        logging.debug(f"Uploaded file '{file.display_name}' as: {file.uri}")
-        file_data : FileData = FileData(file_uri=file.uri, mime_type=mime_type)
-        return Part(file_data=file_data)  
+    async def get_response_from_gemini(self, prompt: str,
+                                 chat: Chat,
+                                 response_mime_type: str = mimetypes.types_map['.txt'],
+                                 base64_decoded: str = None,
+                                 file_paths: list[str] = None,
+                                 ) -> str:
+
+        self.config[self.CONFIG_RESPONSE_MIME_TYPE] = response_mime_type
+        
+        if file_paths:            
+            prompt = prompt + await self.get_json_files_content(file_paths)
+            parts = [Part(text=prompt)]
+        # Handle single image from base64
+        elif base64_decoded:
+            try:
+                image = Image.open(io.BytesIO(base64_decoded))
+                parts.append(Part(inline_data=image))
+            except Exception as e:
+                logging.error(f"Error processing image data: {e}", exc_info=True)
+
+        try:
+            parts = [Part(text=prompt)]
+            response = chat.send_message(message=parts, config=self.config)
+            if response:
+                return response.text
+
+        except Exception as e:
+            logging.error(f"Error using Gemini: {e}", exc_info=True)
+            return f"Sorry, I couldn't process your request with Gemini: {str(e)}"    
+
+    async def get_json_files_content(self, file_paths: list[str]) -> str:
+        combined_content =""
+        for file_path in file_paths:
+            content = await file_utils.read_text_file(file_path)
+            combined_content =  f"{combined_content}\n\n\n{content}"
+        return combined_content
+        
+    
+    def get_json_file_parts(self, file_paths: list[str]) -> list[Part]:
+        parts = []
+        # Upload files
+        for file_path in file_paths:            
+            file = self.upload_large_file_to_google_cloud(file_path)            
+            if file:
+                file_data = self.get_file_data_from_uploaded_files(file, file_path)
+                parts.append(Part(file_data=file_data))
+            else:
+                logging.warning(f"Skipping file that failed to upload: {file_path}")        
+
+        return parts      
+
+    def upload_large_file_to_google_cloud(self, file_path: str) -> File:
+        logging.debug(f"Uploading file: {file_path}")
+        try:
+            file: File = self.gemini_client.files.upload(file=file_path)       
+            logging.debug(f"Successfully uploaded '{file.display_name}' as: {file.uri}")
+            return file      
+        except Exception as e:
+            logging.error(f"Couldn't upload file {file_path}: {e}", exc_info=True)
+            return None
+            
+    def get_file_data_from_uploaded_files(self, file: File, file_path: str) -> FileData:        
+        mime_type = mimetypes.guess_type(file_path)[0] or 'application/octet-stream'
+        if file and file.uri:            
+            logging.debug(f"Uploaded file '{file.display_name}' as: {file.uri}")
+            file_data: FileData = FileData(file_uri=file.uri, mime_type=mime_type)
+            return file_data
+        
+        return None
     
     def delete_all_files(self):
         """
@@ -70,8 +111,14 @@ class GeminiUtils:
         Use with caution in shared environments.
         Consider using selective deletion instead.
         """
-        for f in self.gemini_client.files.list():
+        if self.gemini_client.files:
             try:
-                self.gemini_client.files.delete(name=f.name)
-            except Exception as ex:
-                logging.error(f"Error deleting file {f.name}: {ex}", exc_info=True)
+                for f in self.gemini_client.files.list():
+                    try:
+                        self.gemini_client.files.delete(name=f.name)
+                    except Exception as ex:
+                        logging.error(f"Error deleting file {f.name}: {ex}", exc_info=True)
+            except Exception as e:
+                logging.error(f"Error listing files: {e}", exc_info=True)
+
+
